cleaned up
The scoped css was still an issue so I just kept the global App.css. Different page components are now stored in the "pages" folder and only child components such as "Map" are kept in the "components" folder.
