+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "babel-plugin-react-scoped-css": "^1.1.1",
     "bootstrap": "^5.2.3",
+    "classnames": "^2.3.2",
     "leaflet": "^1.9.3",
     "p5": "^1.5.0",
     "react": "^18.2.0",
