--- src/components/Detail.css
@@ -1,115 +1,0 @@
-body {
-    background-color:  #181D27;
-    font-family: 'Noto Sans', sans-serif;
-    color: white;
-}
-
-p {
-    font-size: 20px;
-}
-
-h1 {
-    text-align: center;
-    margin: 50px;
-    font-size: 50px;
-    font-weight: bolder;
-}
-
-h3 {
-    text-align: left;
-    margin: 50px;
-    font-size: 30px;
-}
-
-table {
-    font-size: 15px;
-    /* margin-top: 10px; */
-}
-
-table thead {
-    position: sticky;
-    top: 0px;
-    margin: 0 0 0 0;
-}
-
-.p5Container {
-    position: absolute;
-    left: 0px;
-    top: 0px;
-    z-index: -1;
-}
-
-#MinorDescription {
-    position: absolute; 
-    color: white;
-}
-
-.tableContainer {
-    background-color: #8595de;
-    border-radius: 10px;
-    height: 50vh;
-    /* padding: 0 10px 10px 10px; */
-    overflow: scroll;
-    overflow: hidden;
-}
-
-.leaflet-container {
-    position: relative;
-    margin: auto;
-    width: 500px;
-    height: 500px;
-    z-index: 1;
-}
-
-.img {
-    margin: auto;
-    display: inline-block;
-}
-
-.img-fluid {
-    max-width: 80%;
-    height: auto;
-}
-
-.title {
-    margin-top: -5vh;
-    padding: 20px;
-}
-
-.top {
-    margin-top: 10vh;
-    padding: 20px;
-}
-
-.vertical-center {
-    margin: 0;
-    position: absolute;
-    top: 50%;
-    -ms-transform: translateY(-50%);
-    transform: translateY(-50%);
-}
-
-.btn {
-    color: white;
-    background-color: #5e6caa;
-    font-size: 30px;
-    font-weight: bold;
-    border-radius: 7px;
-}
-
-.btn-layout {
-    padding: 10px;
-}
-
-.btn:hover {
-    color: #5e6caa;
-    background-color: white;
-}
-
-/* https://www.geeksforgeeks.org/how-to-add-a-pressed-effect-on-button-click-in-css/ */
-.btn:active {
-    transform: scale(0.98);
-    /* Scaling button to 0.98 to its original size */
-    box-shadow: 3px 2px 22px 1px rgba(0, 0, 0, 0.24);
-    /* Lowering the shadow */
-}
