+++ src/pages/Career.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
-import './Detail.css'
 import { MAJOR_DATA_SAMPLE } from "../constants";
-import Map from "./Map";
 
 
 /* Table Generating Function
@@ -10,6 +8,8 @@ import Map from "./Map";
     1. Takes in data about majors and their campus from the "constants.js" file.
     2. Sets the table header and 3 columns. 
     3. Iterates through the data to populate table with rows.
+
+    Can separate this into its own component later on if needed. 
 
 */
 
@@ -20,8 +20,8 @@ const Table = ({data}) => {
             <thead>
                 <tr>
                 <th scope="col">#</th>
-                <th scope="col">Major</th>
-                <th scope="col">Campus</th>
+                <th scope="col">Career</th>
+                <th scope="col">O'Net Link</th>
                 </tr>
             </thead>
             <tbody>
@@ -48,28 +48,33 @@ Contains the overall layout of the detail page and calls other nested functions/
 */
 
 
-class Major extends React.Component{
+class Career extends React.Component{
     constructor() {
         super();
-        this.state = { checked: false };
-        this.handleChange = this.handleChange.bind(this);
     }
-
-    handleChange(checked) {
-        this.setState({ checked });
-    }
-
     render() {
         return (
-            <div className="container-fluid major">
+            <div className="container-fluid">
                 <div className="row">
                     <div className="col-sm-6 title">
                         <h1>{this.props.minor}</h1>
-                        <Table data={ MAJOR_DATA_SAMPLE }/>
                     </div>
                     <div className="col-sm-6 title">
-                        <h1>Placeholder</h1>
-                        <Map />
+                        <h1>List of Careers</h1>
+                    </div>
+                </div>
+                <div className="row">
+                    <div className="col-sm-6 title">
+                        {/* eventually all the sample information shown below will be passed down as props or imported */}
+                        <img className="img-fluid rounded mx-auto d-block" src="src\assets\pexels-mario-schafer-11322619.jpg" alt="Chefs rolling out dough" />
+                        <br /><br />
+                        <div className="row">
+                            <h3>Chefs and Head Cooks</h3>
+                            <p>Direct and may participate in the preparation, seasoning, and cooking of salads, soups, fish, meats, vegetables, desserts, or other foods. May plan and price menu items, order supplies, and keep records and accounts.</p>
+                        </div>
+                    </div>
+                    <div className="col-sm-6 title">
+                        <Table data={ MAJOR_DATA_SAMPLE }/>
                     </div>
                 </div>
             </div>
@@ -77,4 +82,4 @@ class Major extends React.Component{
     }
 }
 
-export default Major
+export default Career
