--- src/pages/MainPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-import 'bootstrap/dist/css/bootstrap.min.css';
-
-
-class MainPage extends React.Component {
-    constructor(props)
-}
