+++ src/pages/Major.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
-import './Detail.css'
 import { MAJOR_DATA_SAMPLE } from "../constants";
+import Map from "../components/Map";
 
 
 /* Table Generating Function
@@ -9,8 +9,6 @@ import { MAJOR_DATA_SAMPLE } from "../constants";
     1. Takes in data about majors and their campus from the "constants.js" file.
     2. Sets the table header and 3 columns. 
     3. Iterates through the data to populate table with rows.
-
-    Can separate this into its own component later on if needed. 
 
 */
 
@@ -21,8 +19,8 @@ const Table = ({data}) => {
             <thead>
                 <tr>
                 <th scope="col">#</th>
-                <th scope="col">Career</th>
-                <th scope="col">O'Net Link</th>
+                <th scope="col">Major</th>
+                <th scope="col">Campus</th>
                 </tr>
             </thead>
             <tbody>
@@ -49,33 +47,28 @@ Contains the overall layout of the detail page and calls other nested functions/
 */
 
 
-class Career extends React.Component{
+class Major extends React.Component{
     constructor() {
         super();
+        this.state = { checked: false };
+        this.handleChange = this.handleChange.bind(this);
     }
+
+    handleChange(checked) {
+        this.setState({ checked });
+    }
+
     render() {
         return (
-            <div className="container-fluid career">
+            <div className="container-fluid">
                 <div className="row">
                     <div className="col-sm-6 title">
                         <h1>{this.props.minor}</h1>
-                    </div>
-                    <div className="col-sm-6 title">
-                        <h1>List of Careers</h1>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-sm-6 title">
-                        {/* eventually all the sample information shown below will be passed down as props or imported */}
-                        <img className="img-fluid rounded mx-auto d-block" src="src\assets\pexels-mario-schafer-11322619.jpg" alt="Chefs rolling out dough" />
-                        <br /><br />
-                        <div className="row">
-                            <h3>Chefs and Head Cooks</h3>
-                            <p>Direct and may participate in the preparation, seasoning, and cooking of salads, soups, fish, meats, vegetables, desserts, or other foods. May plan and price menu items, order supplies, and keep records and accounts.</p>
-                        </div>
-                    </div>
-                    <div className="col-sm-6 title">
                         <Table data={ MAJOR_DATA_SAMPLE }/>
+                    </div>
+                    <div className="col-sm-6 title">
+                        <h1>Placeholder</h1>
+                        <Map />
                     </div>
                 </div>
             </div>
@@ -83,4 +76,4 @@ class Career extends React.Component{
     }
 }
 
-export default Career
+export default Major
