--- src/pages/detail-ag.js
@@ -1,10 +1,0 @@
-
-import React from 'react';
-  
-const Blogs = () => {
-  return (
-    <h1>You can write your blogs!</h1>
-  );
-};
-  
-export default Blogs;
